requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0
gunicorn==20.1.0
Werkzeug==2.3.6
